Redirect wizard page and rename progress API
