+++ apps/cms/__tests__/dashboardSkip.integration.test.tsx
@@ -79,7 +79,7 @@ beforeEach(() => {
     completed: Object.fromEntries(required.map((s) => [s.id, "complete"])),
   };
   global.fetch = jest.fn((url: string, init?: RequestInit) => {
-    if (url === "/cms/api/wizard-progress") {
+    if (url === "/cms/api/configurator-progress") {
       if (init?.method === "PUT") {
         const body = JSON.parse(init.body as string);
         serverState.state = { ...serverState.state, ...(body.data ?? {}) };
