+++ apps/cms/src/app/cms/wizard/page.tsx
@@ -1,8 +1,10 @@
-// Temporary page used for testing the wizard route. In the real application this
-// route redirects to the configurator, but for unit tests we render a simple
-// placeholder so server rendering can be verified without pulling in the entire
-// configurator implementation.
+// apps/cms/src/app/cms/wizard/page.tsx
+// Legacy wizard route retained for backwards‑compatibility. Requests are
+// redirected to the configurator which replaced the wizard implementation.
 
-export default function WizardPage() {
-  return <h2>Shop Details</h2>;
+import { redirect } from "next/navigation";
+
+export default function WizardPage(): void {
+  redirect("/cms/configurator");
 }
+
