+++ packages/i18n/__tests__/msw.test.ts
@@ -30,22 +30,22 @@ describe('msw default handlers', () => {
     expect(json).toEqual([]);
   });
 
-  test('GET /cms/api/wizard-progress', async () => {
-    const res = await fetch('http://localhost/cms/api/wizard-progress');
+  test('GET /cms/api/configurator-progress', async () => {
+    const res = await fetch('http://localhost/cms/api/configurator-progress');
     expect(res.status).toBe(200);
     const json = await res.json();
     expect(json).toEqual({ state: {}, completed: {} });
   });
 
-  test('PUT /cms/api/wizard-progress', async () => {
-    const res = await fetch('http://localhost/cms/api/wizard-progress', { method: 'PUT' });
+  test('PUT /cms/api/configurator-progress', async () => {
+    const res = await fetch('http://localhost/cms/api/configurator-progress', { method: 'PUT' });
     expect(res.status).toBe(200);
     const json = await res.json();
     expect(json).toEqual({});
   });
 
-  test('PATCH /cms/api/wizard-progress', async () => {
-    const res = await fetch('http://localhost/cms/api/wizard-progress', { method: 'PATCH' });
+  test('PATCH /cms/api/configurator-progress', async () => {
+    const res = await fetch('http://localhost/cms/api/configurator-progress', { method: 'PATCH' });
     expect(res.status).toBe(200);
     const json = await res.json();
     expect(json).toEqual({});
