+++ test/msw/server.ts
@@ -26,13 +26,13 @@ export const handlers = [
   rest.get("/cms/api/page-templates", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json([]))
   ),
-  rest.get("/cms/api/wizard-progress", (_req, res, ctx) =>
+  rest.get("/cms/api/configurator-progress", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json({ state: {}, completed: {} }))
   ),
-  rest.put("/cms/api/wizard-progress", (_req, res, ctx) =>
+  rest.put("/cms/api/configurator-progress", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json({}))
   ),
-  rest.patch("/cms/api/wizard-progress", (_req, res, ctx) =>
+  rest.patch("/cms/api/configurator-progress", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json({}))
   ),
   // Allow API route tests to hit local handlers without mocking
