+++ test/mswServer.ts
@@ -6,13 +6,13 @@ import { setupServer } from "msw/node";
  * Handlers can be added in individual tests via `server.use(...)`.
  */
 const handlers = [
-  rest.get("/cms/api/wizard-progress", (_req, res, ctx) =>
+  rest.get("/cms/api/configurator-progress", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json({ state: {}, completed: {} }))
   ),
-  rest.put("/cms/api/wizard-progress", (_req, res, ctx) =>
+  rest.put("/cms/api/configurator-progress", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json({}))
   ),
-  rest.patch("/cms/api/wizard-progress", (_req, res, ctx) =>
+  rest.patch("/cms/api/configurator-progress", (_req, res, ctx) =>
     res(ctx.status(200), ctx.json({}))
   ),
 ];
